rewrite annotation now a helper on the pattern query
